refactor(plugins): define hook handlers in lifecycle module
